refactor(persistence): promote to directory, split tests (#1266)
persistence.rs (509 LOC) is promoted to a directory module:
- src/persistence/mod.rs: 217 LOC (down from 509)
- src/persistence/tests.rs: new (~292 LOC)

Pure refactor: zero logic changes, all 45 tests pass.

Closes 17/70 of #1266 (this loop iteration), 26 PRs cumulative since #1285.
